refactor: separate data loaders into own files and rewrite MZTab loader
