+++ pyproject.toml
@@ -32,7 +32,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["winnow", "winnow.scripts", "winnow.fdr", "winnow.datasets", "winnow.calibration", "winnow.compat", "winnow.configs", "winnow.configs.data_loader", "winnow.configs.fdr_method", "winnow.utils"]
+packages = ["winnow", "winnow.scripts", "winnow.fdr", "winnow.datasets", "winnow.datasets.data_loaders", "winnow.calibration", "winnow.compat", "winnow.configs", "winnow.configs.data_loader", "winnow.configs.fdr_method", "winnow.utils"]
 
 [tool.setuptools.package-data]
 winnow = ["configs/**/*.yaml", "configs/**/*.yml"]
